Allow singular `inputArtifact` in stage contexts
inputArtifact used to be a shortcut for a single-element list of
inputArtifacts, but this is not compatible with the Kustomize version of
the Bake (Manifest) stage. This change allows for inputArtifact or
inputArtifacts, but rejects the presence of both. It leaves
inputArtifact (singular) as a single object rather than wrapping it.