fix(base64): Base64 artifacts not utilizing encode flag
This commit addresses an issue where Base64ArtifactDefinition was not
properly utilizing the `shouldEncode` method/field. This was mainly due
to having multiple `@Builder`s, and also `Boolean`s defaulting to false.
To fix this we simply add a `null` check against the new flag that was
added

* update(gitignore): add some appropriate gitignore options

* idea folder
* lombok.config

Authored by: benjamin-j-powell <bjp@apple.com>